backend/requirements.txt
Flask==3.0.0
python-dotenv==1.0.1
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.6.0
bcrypt>=4.2.0,<5

